feat(ui): add ability to disable game option tabs
Introduces a `disabled_items` prop to `GameOptionTabs` component, allowing specific tabs to be marked as disabled. Disabled tabs are visually indicated with greyed-out text, a strikethrough, and a `not-allowed` cursor, and are no longer clickable.
This feature is immediately utilized in `GameOptions.tsx` to disable certain game modes and item options that are not yet available.

refactor(auth): change token fetch console.error to console.warn

Adjusts the logging severity for "No refresh token found." and "Token not found in response data" messages in `refreshAccessToken` function from `console.error` to `console.warn`.
This change reflects that these scenarios are not critical application errors, but rather expe